feat(security): 调整security后台模块

+++ mk-security/security/src/main/java/com/github/xujiaji/mk/security/service/impl/MkSecUserServiceImpl.java
@@ -1,20 +1,13 @@
 package com.github.xujiaji.mk.security.service.impl;
 
-import cn.hutool.core.bean.BeanUtil;
 import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.baomidou.mybatisplus.core.metadata.IPage;
-import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.github.xujiaji.mk.common.base.BaseServiceImpl;
 import com.github.xujiaji.mk.common.base.Consts;
-import com.github.xujiaji.mk.common.entity.MkUser;
 import com.github.xujiaji.mk.common.exception.RequestActionException;
-import com.github.xujiaji.mk.common.service.IUserInfoService;
 import com.github.xujiaji.mk.common.service.IUserLoginLogService;
 import com.github.xujiaji.mk.common.util.UserUtil;
-import com.github.xujiaji.mk.common.vo.PageVO;
 import com.github.xujiaji.mk.file.service.impl.MkFileServiceImpl;
-import com.github.xujiaji.mk.security.entity.MkAdminUser;
 import com.github.xujiaji.mk.security.entity.MkSecUser;
 import com.github.xujiaji.mk.security.mapper.MkSecRoleMapper;
 import com.github.xujiaji.mk.security.mapper.MkSecUserMapper;
@@ -24,9 +17,8 @@ import com.github.xujiaji.mk.security.playload.AdminEditCondition;
 import com.github.xujiaji.mk.security.playload.AdminLoginCondition;
 import com.github.xujiaji.mk.security.service.IMkSecUserService;
 import com.github.xujiaji.mk.security.util.JwtUtil;
-import com.github.xujiaji.mk.security.util.SecurityUtil;
 import com.github.xujiaji.mk.security.vo.AdminLoginSuccessVO;
-import com.github.xujiaji.mk.security.vo.UserPrincipal;
+import com.github.xujiaji.mk.security.vo.MkSecUserDetails;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -38,9 +30,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpServletRequest;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
 
 /**
  * <p>
@@ -54,7 +43,6 @@ import java.util.stream.Collectors;
 @Service
 public class MkSecUserServiceImpl extends BaseServiceImpl<MkSecUserMapper, MkSecUser> implements IMkSecUserService {
 
-    private final IUserInfoService userInfoService;
     private final MkSecUserRoleMapper secUserRoleMapper;
     private final MkSecRoleMapper secRoleMapper;
     private final AuthenticationManager authenticationManager;
@@ -67,35 +55,6 @@ public class MkSecUserServiceImpl extends BaseServiceImpl<MkSecUserMapper, MkSec
     private IUserLoginLogService userLoginLogService;
 
     @Override
-    public PageVO<MkAdminUser> adminUserPage(Page<MkSecUser> page) {
-        IPage<MkSecUser> secUserPage = page(page);
-        List<MkAdminUser> list = new ArrayList<>();
-        val userDetailsList = userInfoService.getUserDetailsList(secUserPage.getRecords().stream().map(MkSecUser::getUserId).collect(Collectors.toList()));
-
-        for (MkSecUser record : secUserPage.getRecords()) {
-            val mkUser = userDetailsList.stream().filter(f -> f.getId().equals(record.getUserId())).findFirst().orElseGet(MkUser::new);
-            val mkAdminUser = BeanUtil.copyProperties(mkUser, MkAdminUser.class);
-            mkAdminUser.setPassword(null);
-            mkAdminUser.setStatus(record.getStatus());
-            mkAdminUser.setSecUserId(record.getId());
-            mkAdminUser.setRole(SecurityUtil.getCurrentUser().getRoles().stream().reduce((s1, s2) -> s1 + "," + s2).get());
-            mkAdminUser.setLoginLog(userInfoService.lastLoginLogBy(record.getUserId()));
-            if (mkUser.getAvatar() != null) {
-                mkAdminUser.setAvatar(fileService.getPathById(mkUser.getAvatar()));
-            }
-            list.add(mkAdminUser);
-        }
-
-        val mkUserPageVO = new PageVO<MkAdminUser>();
-        mkUserPageVO.setPages(secUserPage.getPages());
-        mkUserPageVO.setList(list);
-        mkUserPageVO.setPage(secUserPage.getCurrent());
-        mkUserPageVO.setTotal(secUserPage.getTotal());
-        mkUserPageVO.setSize(secUserPage.getSize());
-        return mkUserPageVO;
-    }
-
-    @Override
     @Transactional(rollbackFor = Exception.class)
     public void adminAdd(AdminAddCondition request) {
         // 判断是否有这个角色
@@ -104,19 +63,14 @@ public class MkSecUserServiceImpl extends BaseServiceImpl<MkSecUserMapper, MkSec
             throw new RequestActionException("没有这个角色！你选的啥哦～");
         }
 
-        MkUser mkUser = userInfoService.getUserByUsername(request.getUsername());
-
-        if (mkUser == null) { // 这个用户不存在，那么创建这个用户
-            mkUser = userInfoService.createUserByUsername(request.getUsername(), request.getPassword(), request.getAvatarId());
-        }
-
-        if (baseMapper.selectCount(new QueryWrapper<MkSecUser>().eq("user_id", mkUser.getId())) > 0) {
-            throw new RequestActionException("该用户已经就是管理员了");
+        if (baseMapper.selectCount(new QueryWrapper<MkSecUser>().eq("username", request.getUsername())) > 0) {
+            throw new RequestActionException("已添加过该管理员了");
         }
 
         val mkSecUser = new MkSecUser();
         mkSecUser.setStatus(Consts.ENABLE);
-        mkSecUser.setUserId(mkUser.getId());
+        mkSecUser.setUsername(request.getUsername());
+        mkSecUser.setPassword(passwordService.encode(request.getPassword()));
         checkInsertSuccess(baseMapper.insert(mkSecUser));
 
         checkInsertSuccess(secUserRoleMapper.addSecUserRole(mkSecUser.getId(), request.getRoleId()));
@@ -125,31 +79,38 @@ public class MkSecUserServiceImpl extends BaseServiceImpl<MkSecUserMapper, MkSec
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void adminEdit(AdminEditCondition request) {
-        // 判断是否有这个角色
-        val mkSecRole = secRoleMapper.selectById(request.getRoleId());
-        if (mkSecRole == null) {
-            throw new RequestActionException("没有这个角色！你选的啥哦～");
+
+        val mkSecUser = baseMapper.selectById(request.getId());
+        if (mkSecUser == null) {
+            throw new RequestActionException("没有这个管理员无法编辑");
         }
-        MkUser user = userInfoService.getUserDetails(userUtil.currentUserIdNotNull());
-        // 如果用户名不相等，那么改用户名
-        if (!user.getUsername().equals(request.getUsername())) {
-            MkUser mkUser = userInfoService.getUserByUsername(request.getUsername());
-            if (mkUser != null) {
+
+        if (request.getRoleId() != null) {
+            // 判断是否有这个角色
+            val mkSecRole = secRoleMapper.selectById(request.getRoleId());
+            if (mkSecRole == null) {
+                throw new RequestActionException("没有这个角色！你选的啥哦～");
+            }
+            // 先删除之前的角色
+            checkDeleteSuccess(secUserRoleMapper.deleteBySecUserId(mkSecUser.getId()));
+            // 再添加现在的角色
+            checkInsertSuccess(secUserRoleMapper.addSecUserRole(mkSecUser.getId(), request.getRoleId()));
+        }
+
+        if (StrUtil.isNotBlank(request.getUsername())) {
+            if (baseMapper.selectCount(new QueryWrapper<MkSecUser>().eq("username", request.getUsername())) > 0) {
                 throw new RequestActionException("该用户名已被占用");
             }
-            user.setUsername(request.getUsername());
+            mkSecUser.setUsername(request.getUsername());
         }
-        user.setNickname(request.getNickname());
-        user.setAvatar(request.getAvatarId());
-        val mkSecUser = baseMapper.selectByUserId(user.getId());
-        // 先删除之前的角色
-        checkDeleteSuccess(secUserRoleMapper.deleteBySecUserId(mkSecUser.getId()));
-        // 再添加现在的角色
-        checkInsertSuccess(secUserRoleMapper.addSecUserRole(mkSecUser.getId(), request.getRoleId()));
+
+        mkSecUser.setNickname(request.getNickname());
+        mkSecUser.setAvatar(request.getAvatarId());
+
         if (StrUtil.isNotBlank(request.getPassword())) {
-            user.setPassword(passwordService.encode(request.getPassword()));
+            mkSecUser.setPassword(passwordService.encode(request.getPassword()));
         }
-        userInfoService.updateUser(user);
+        checkUpdateSuccess(updateById(mkSecUser));
     }
 
     @Override
@@ -168,10 +129,10 @@ public class MkSecUserServiceImpl extends BaseServiceImpl<MkSecUserMapper, MkSec
 
         String jwt = jwtUtil.createJWT(authentication, request.getRememberMe());
 
-        val principal = (UserPrincipal) authentication.getPrincipal();
+        val principal = (MkSecUserDetails) authentication.getPrincipal();
 
         if (userLoginLogService != null) {
-            userLoginLogService.insertLog(principal.getUserId(), Consts.LoginType.USERNAME, hsr);
+            userLoginLogService.insertLog(principal.getId(), Consts.LoginType.USERNAME, hsr);
         }
 
         return AdminLoginSuccessVO.builder()
